% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{action_btn}
\alias{action_btn}
\title{Action Button (Styled)}
\usage{
action_btn(id, label, icon = NULL, class = "btn-primary")
}
\arguments{
\item{id}{Input ID.}

\item{label}{Button label.}

\item{icon}{An optional icon.}

\item{class}{CSS class. Defaults to "btn-primary".}
}
\description{
A styled action button consistent with the v3 theme.
}
\keyword{internal}
