% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascseurat-package.R
\docType{package}
\name{ascseurat-package}
\alias{ascseurat}
\alias{ascseurat-package}
\title{ascseurat: Analytical Single Cell Seurat-Based Web Application}
\description{
Asc-Seurat (Analytical Single Cell Seurat-Based Web Application) is an interactive Shiny application for the analysis of single-cell RNA sequencing (scRNA-seq) data. It provides a guided, step-by-step interface for quality control, normalization, clustering, differential expression, trajectory inference, and cell-type annotation, all powered by the Seurat v5 framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Pereira-Lab-UF/asc_seurat}
  \item \url{https://pereira-lab-uf.github.io/asc_seurat/}
  \item Report bugs at \url{https://github.com/Pereira-Lab-UF/asc_seurat/issues}
}

}
\author{
\strong{Maintainer}: Felipe Marques de Almeida \email{almeidafmarques@outlook.com}

Authors:
\itemize{
  \item Felipe Marques de Almeida \email{almeidafmarques@outlook.com}
  \item Wendell Pereira
  \item Marcelo Kirst
}

}
\keyword{internal}
