% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compat.R
\name{detect_input_format}
\alias{detect_input_format}
\title{Detect Input Data Format}
\usage{
detect_input_format(path)
}
\arguments{
\item{path}{Character. Path to the input file or directory.}
}
\value{
Character. One of: "10x_dir", "10x_h5", "h5ad", "csv", "tsv", "rds", "unknown".
}
\description{
Determines the format of input data (10X directory, h5, h5ad, csv, rds)
and returns a standardized format identifier.
}
\keyword{internal}
