% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compat.R
\name{format_bytes}
\alias{format_bytes}
\title{Format Bytes for Human Readability}
\usage{
format_bytes(bytes)
}
\arguments{
\item{bytes}{Numeric scalar in bytes.}
}
\value{
Character string.
}
\description{
Format Bytes for Human Readability
}
\keyword{internal}
