% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{input_de_test}
\alias{input_de_test}
\title{Select Statistical Test for DE}
\usage{
input_de_test(id)
}
\arguments{
\item{id}{Input ID.}
}
\description{
Select Statistical Test for DE
}
\keyword{internal}
