% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{input_markers_file}
\alias{input_markers_file}
\title{File Input for Markers List}
\usage{
input_markers_file(id)
}
\arguments{
\item{id}{Input ID.}
}
\description{
File Input for Markers List
}
\keyword{internal}
