% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{input_mito_regex}
\alias{input_mito_regex}
\title{Mitochondrial Gene Regex Input}
\usage{
input_mito_regex(id, value = "")
}
\arguments{
\item{id}{Input ID.}

\item{value}{Default regex pattern.}
}
\description{
Mitochondrial Gene Regex Input
}
\keyword{internal}
