% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{input_norm_method}
\alias{input_norm_method}
\title{Select Normalization Method}
\usage{
input_norm_method(id)
}
\arguments{
\item{id}{Input ID.}
}
\description{
Select Normalization Method
}
\keyword{internal}
