% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{input_plot_format}
\alias{input_plot_format}
\title{Select Input for Plot Format}
\usage{
input_plot_format(id)
}
\arguments{
\item{id}{Input ID.}
}
\description{
Select Input for Plot Format
}
\keyword{internal}
