% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{input_plot_height}
\alias{input_plot_height}
\title{Numeric Input for Plot Height}
\usage{
input_plot_height(id, value = 15)
}
\arguments{
\item{id}{Input ID.}

\item{value}{Default value in cm.}
}
\description{
Numeric Input for Plot Height
}
\keyword{internal}
