% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{input_plot_res}
\alias{input_plot_res}
\title{Select Input for Plot Resolution}
\usage{
input_plot_res(id)
}
\arguments{
\item{id}{Input ID.}
}
\description{
Select Input for Plot Resolution
}
\keyword{internal}
