% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{input_plot_width}
\alias{input_plot_width}
\title{Numeric Input for Plot Width}
\usage{
input_plot_width(id, value = 15)
}
\arguments{
\item{id}{Input ID.}

\item{value}{Default value in cm.}
}
\description{
Numeric Input for Plot Width
}
\keyword{internal}
