% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_advanced_plots.R
\name{mod_advanced_plots_server}
\alias{mod_advanced_plots_server}
\title{Advanced Plots Module — Server}
\usage{
mod_advanced_plots_server(
  id,
  seurat_obj_single = NULL,
  seurat_obj_multi = NULL
)
}
\arguments{
\item{id}{Module namespace ID.}

\item{seurat_obj_single}{Reactive Seurat object from single-sample pipeline.}

\item{seurat_obj_multi}{Reactive Seurat object from integration pipeline.}
}
\description{
Advanced Plots Module — Server
}
\keyword{internal}
