% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_advanced_plots.R
\name{mod_advanced_plots_ui}
\alias{mod_advanced_plots_ui}
\title{Advanced Plots Module — UI}
\usage{
mod_advanced_plots_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Advanced Plots Module — UI
}
\keyword{internal}
