% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_celltype_annotation.R
\name{mod_celltype_annotation_server}
\alias{mod_celltype_annotation_server}
\title{Automated Cell-Type Annotation Module - Server}
\usage{
mod_celltype_annotation_server(
  id,
  seurat_obj_single = NULL,
  seurat_obj_multi = NULL
)
}
\arguments{
\item{id}{Module namespace ID.}

\item{seurat_obj_single}{Optional reactive Seurat object from the single-sample workflow.}

\item{seurat_obj_multi}{Optional reactive Seurat object from the integration workflow.}
}
\description{
Automated Cell-Type Annotation Module - Server
}
\keyword{internal}
