% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_celltype_annotation.R
\name{mod_celltype_annotation_ui}
\alias{mod_celltype_annotation_ui}
\title{Automated Cell-Type Annotation Module - UI}
\usage{
mod_celltype_annotation_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Automated Cell-Type Annotation Module - UI
}
\keyword{internal}
