% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clustering.R
\name{mod_clustering_server}
\alias{mod_clustering_server}
\title{Clustering Module - Server}
\usage{
mod_clustering_server(id, norm_result)
}
\arguments{
\item{id}{Module namespace ID.}

\item{norm_result}{Reactive Seurat object or a list of reactives returned by
the normalization module.}
}
\value{
Reactive clustered Seurat object.
}
\description{
Clustering Module - Server
}
\keyword{internal}
