% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clustering.R
\name{mod_clustering_ui}
\alias{mod_clustering_ui}
\title{Clustering Module - UI}
\usage{
mod_clustering_ui(
  id,
  step_number = 4,
  download_label = "Download Clustered Object"
)
}
\arguments{
\item{id}{Module namespace ID.}

\item{step_number}{Numeric step number to show in the workflow title.}

\item{download_label}{Character label for the clustered object download.}
}
\description{
Clustering Module - UI
}
\keyword{internal}
