% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loading.R
\name{mod_data_loading_server}
\alias{mod_data_loading_server}
\title{Data Loading Module — Server}
\usage{
mod_data_loading_server(id, trigger_demo = NULL)
}
\arguments{
\item{id}{Module namespace ID.}

\item{trigger_demo}{Optional reactive value. When incremented above 0, demo
data loads automatically (used by the Demo tab).}
}
\value{
Reactive Seurat object.
}
\description{
Data Loading Module — Server
}
\keyword{internal}
