% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loading.R
\name{mod_data_loading_ui}
\alias{mod_data_loading_ui}
\title{Data Loading Module — UI}
\usage{
mod_data_loading_ui(id, demo_mode = FALSE)
}
\arguments{
\item{id}{Module namespace ID.}

\item{demo_mode}{Logical. When TRUE, hides input controls (used by Demo tab).}
}
\description{
Data Loading Module — UI
}
\keyword{internal}
