% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_de.R
\name{mod_de_server}
\alias{mod_de_server}
\title{Differential Expression Module — Server}
\usage{
mod_de_server(id, seurat_obj, allow_conserved = FALSE)
}
\arguments{
\item{id}{Module namespace ID.}

\item{seurat_obj}{Reactive clustered Seurat object.}

\item{allow_conserved}{Logical. Whether to expose conserved-marker analysis.}
}
\description{
Differential Expression Module — Server
}
\keyword{internal}
