% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_de.R
\name{mod_de_ui}
\alias{mod_de_ui}
\title{Differential Expression Module — UI}
\usage{
mod_de_ui(id, allow_conserved = FALSE, step_number = 5)
}
\arguments{
\item{id}{Module namespace ID.}

\item{allow_conserved}{Logical. Whether to expose conserved-marker analysis.}

\item{step_number}{Numeric step number to show in the workflow title.}
}
\description{
Differential Expression Module — UI
}
\keyword{internal}
