% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_integration.R
\name{mod_integration_server}
\alias{mod_integration_server}
\title{Integration Module — Server}
\usage{
mod_integration_server(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
Reactive integrated Seurat object.
}
\description{
Integration Module — Server
}
\keyword{internal}
