% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_integration.R
\name{mod_integration_ui}
\alias{mod_integration_ui}
\title{Integration Module — UI}
\usage{
mod_integration_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Integration Module — UI
}
\keyword{internal}
