% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normalization.R
\name{mod_normalization_server}
\alias{mod_normalization_server}
\title{Normalization Module - Server}
\usage{
mod_normalization_server(id, seurat_obj)
}
\arguments{
\item{id}{Module namespace ID.}

\item{seurat_obj}{Reactive Seurat object from QC module.}
}
\value{
A list of reactives with the normalized Seurat object and chosen PCs.
}
\description{
Normalization Module - Server
}
\keyword{internal}
