% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_normalization.R
\name{mod_normalization_ui}
\alias{mod_normalization_ui}
\title{Normalization Module - UI}
\usage{
mod_normalization_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Normalization Module - UI
}
\keyword{internal}
