% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qc.R
\name{mod_qc_server}
\alias{mod_qc_server}
\title{QC Module — Server}
\usage{
mod_qc_server(id, seurat_obj)
}
\arguments{
\item{id}{Module namespace ID.}

\item{seurat_obj}{Reactive Seurat object from data loading module.}
}
\value{
Reactive filtered Seurat object.
}
\description{
QC Module — Server
}
\keyword{internal}
