% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trajectory.R
\name{mod_trajectory_server}
\alias{mod_trajectory_server}
\title{Trajectory Inference Module - Server}
\usage{
mod_trajectory_server(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Trajectory Inference Module - Server
}
\keyword{internal}
