% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trajectory.R
\name{mod_trajectory_ui}
\alias{mod_trajectory_ui}
\title{Trajectory Inference Module - UI}
\usage{
mod_trajectory_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Trajectory Inference Module - UI
}
\keyword{internal}
