% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visualization.R
\name{mod_visualization_ui}
\alias{mod_visualization_ui}
\title{Visualization Module — UI}
\usage{
mod_visualization_ui(id, step_number = 6)
}
\arguments{
\item{id}{Module namespace ID.}

\item{step_number}{Numeric step number to show in the workflow title.}
}
\description{
Visualization Module — UI
}
\keyword{internal}
