% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{paga_available}
\alias{paga_available}
\title{Check if PAGA is Available}
\usage{
paga_available()
}
\value{
Logical. TRUE if Python and scanpy are available.
}
\description{
Check if PAGA is Available
}
\keyword{internal}
