% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{preserve_file_input_scroll_js}
\alias{preserve_file_input_scroll_js}
\title{Preserve Scroll Position Around File Input Dialogs}
\usage{
preserve_file_input_scroll_js(id)
}
\arguments{
\item{id}{Input ID.}
}
\description{
Preserve Scroll Position Around File Input Dialogs
}
\keyword{internal}
