% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trajectory.R
\name{run_monocle3_trajectory}
\alias{run_monocle3_trajectory}
\title{Run Monocle 3 Trajectory Inference}
\usage{
run_monocle3_trajectory(obj, start_cluster = NULL, end_cluster = NULL)
}
\description{
Run Monocle 3 Trajectory Inference
}
\keyword{internal}
