% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trajectory.R
\name{run_paga_trajectory}
\alias{run_paga_trajectory}
\title{Run PAGA Trajectory Inference}
\usage{
run_paga_trajectory(obj, start_cluster = NULL, end_cluster = NULL)
}
\description{
Run PAGA Trajectory Inference
}
\keyword{internal}
