% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trajectory.R
\name{run_pseudotimede_fast}
\alias{run_pseudotimede_fast}
\title{Run PseudotimeDE-fast on Slingshot Output}
\usage{
run_pseudotimede_fast(sce, gene_subset = NULL, cores = 2)
}
\arguments{
\item{sce}{A SingleCellExperiment with trajectory_path and trajectory_pseudotime columns.}

\item{gene_subset}{Optional character vector of genes to test.}

\item{cores}{Number of workers to use.}
}
\value{
A list containing the engine name and result table.
}
\description{
Run PseudotimeDE-fast on Slingshot Output
}
\keyword{internal}
