% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_seurat.R
\name{run_qc_filter}
\alias{run_qc_filter}
\title{Run QC on a Seurat Object}
\usage{
run_qc_filter(
  obj,
  mito_pattern = "^MT-",
  min_features = 200,
  max_features = Inf,
  max_mito_pct = 5
)
}
\arguments{
\item{obj}{A Seurat object.}

\item{mito_pattern}{Character. Regex to identify mitochondrial genes.}

\item{min_features}{Integer. Min genes per cell.}

\item{max_features}{Integer. Max genes per cell.}

\item{max_mito_pct}{Numeric. Max mitochondrial percentage.}
}
\value{
A filtered Seurat object.
}
\description{
Adds percentage mitochondrial content and filters cells.
}
\keyword{internal}
