% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trajectory.R
\name{run_slingshot_trajectory}
\alias{run_slingshot_trajectory}
\title{Run Slingshot Trajectory Inference}
\usage{
run_slingshot_trajectory(obj, start_cluster = NULL, end_cluster = NULL)
}
\description{
Run Slingshot Trajectory Inference
}
\keyword{internal}
