% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trajectory.R
\name{run_tradeseq}
\alias{run_tradeseq}
\title{Run tradeSeq on Slingshot Output}
\usage{
run_tradeseq(sce, assignment, gene_subset = NULL, nknots = 6)
}
\arguments{
\item{sce}{A SingleCellExperiment with Slingshot results.}

\item{assignment}{Slingshot assignment metadata from extract_slingshot_assignments().}

\item{gene_subset}{Optional character vector of genes to test.}

\item{nknots}{Number of knots for tradeSeq GAM fitting.}
}
\value{
A list containing the fitted tradeSeq object and a result table.
}
\description{
Run tradeSeq on Slingshot Output
}
\keyword{internal}
