% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{with_spinner}
\alias{with_spinner}
\title{Custom Spinner Wrapper}
\usage{
with_spinner(ui_element, hide_ui = FALSE)
}
\arguments{
\item{ui_element}{A Shiny UI element.}

\item{hide_ui}{Logical. Whether to hide the UI while loading.}
}
\value{
The UI element wrapped with a spinner.
}
\description{
Wraps a UI element with a loading spinner.
}
\keyword{internal}
